+++ include/DetectorWithAnalyticNonGaussianPSF.h
@@ -24,7 +24,7 @@ class DetectorWithAnalyticNonGaussianPSF: public Detector
 {
     public:
 
-        DetectorWithAnalyticNonGaussianPSF(ConfigurationParameters &configParam, HDF5File &hdf5File, Camera &camera, TemperatureGenerator &feeTemperatureGenerator, TemperatureGenerator &detectorTemperatureGenerator);
+        DetectorWithAnalyticNonGaussianPSF(ConfigurationParameters &configParam, HDF5File &hdf5File, Camera &camera, TemperatureGenerator &feeTemperatureGenerator, TemperatureGenerator &detectorTemperatureGenerator, double readoutTimeBeforeNextExposure, double readoutTimeDuringNextExposure);
         virtual ~DetectorWithAnalyticNonGaussianPSF();
 
         virtual double takeExposure(int exposureNr, double startTime, double exposureTime) override;