+++ source/Detector.cpp
@@ -139,9 +139,10 @@ double IntegralOfAnalyticSignalResponse::operator()(unsigned i, unsigned j, bool
  * \param configParam    Configuration parameters for the detector.
  * \param hdf5file       HFD5 file to write the detector images to.
  * \param camera         Camera to which to attach the detector.
+ * \param readoutTimeBeforeNextExposure Duration of the readout that takes place before the next exposure can start.
  */
 
-Detector::Detector(ConfigurationParameters &configParam, HDF5File &hdf5file, Camera &camera, TemperatureGenerator &feeTemperatureGenerator, TemperatureGenerator &detectorTemperatureGenerator)
+Detector::Detector(ConfigurationParameters &configParam, HDF5File &hdf5file, Camera &camera, TemperatureGenerator &feeTemperatureGenerator, TemperatureGenerator &detectorTemperatureGenerator, double readoutTimeBeforeNextExposure, double readoutTimeDuringNextExposure)
 : HDF5Writer(hdf5file),
   includeCosmicsInSubField(true), includeCosmicsInSmearingMap(true), includeCosmicsInBiasMap(true),
   includeBFE(true),
@@ -163,6 +164,9 @@ Detector::Detector(ConfigurationParameters &configParam, HDF5File &hdf5file, Cam
 
     configure(configParam);
 
+    this->readoutTimeBeforeNextExposure = readoutTimeBeforeNextExposure;
+    this->readoutTimeDuringNextExposure = readoutTimeDuringNextExposure;
+
     // Create the groups in the HDF5 file where the different maps (i.e. pixel map,
     // bias register map, smearing map, etc.) will be saved. This needs to be done
     // BEFORE other methods write arrays to HDF5.
@@ -176,7 +180,9 @@ Detector::Detector(ConfigurationParameters &configParam, HDF5File &hdf5file, Cam
     // Allocate memory for the different maps
 
     pixelMap.zeros(numRowsPixelMap, numColumnsPixelMap);
-    biasMap.zeros(numRowsBiasMap, numColumnsPixelMap);
+    biasMapLeft.zeros(numRowsBiasMap, numColumnsBiasMap);
+    biasMapRight.zeros(numRowsBiasMap, numColumnsBiasMap);
+
     smearingMap.zeros(numRowsSmearingMap, numColumnsPixelMap);
     throughputMap.ones(numRowsPixelMap, numColumnsPixelMap);
 
@@ -255,7 +261,7 @@ void Detector::updateParameters(double time)
 /**
  * \brief Configure the Detector object using the ConfigurationParameters
  * 
- * \param configParam: the configuration parameters 
+ * \param configParam: Configuration parameters
  **/
 
  void Detector::configure(ConfigurationParameters &configParam)
@@ -283,9 +289,9 @@ void Detector::updateParameters(double time)
         numRows               = configParam.getIntegerAt("CCDPositions/NumRows", idx);
         numColumns            = configParam.getIntegerAt("CCDPositions/NumColumns", idx);
 
-        string groupID        = configParam.getString("Telescope/GroupID");
+        isFastCamera          = configParam.getString("Telescope/GroupID") == "Fast";
         
-        if (groupID == "Fast")
+        if (isFastCamera)
         {
             firstRowExposed       = configParam.getIntegerAt("CCDPositions/FirstRowForFastCamera", idx);
         }
@@ -324,7 +330,6 @@ void Detector::updateParameters(double time)
     fullWellSaturationLimit             = configParam.getLong("CCD/FullWellSaturation");
     digitalSaturationLimit              = configParam.getLong("CCD/DigitalSaturation");
     readoutNoise                        = configParam.getDouble("CCD/ReadoutNoise");
-    readoutTime                         = configParam.getDouble("CCD/ReadoutTime");
     expectedValueVignetting             = configParam.getDouble("CCD/Vignetting/ExpectedValue");
     particulateContaminationEfficiency  = configParam.getDouble("CCD/Contamination/ParticulateContaminationEfficiency");
     molecularContaminationEfficiency    = configParam.getDouble("CCD/Contamination/MolecularContaminationEfficiency");
@@ -333,6 +338,23 @@ void Detector::updateParameters(double time)
     refValueGainRight         = configParam.getDouble("CCD/Gain/RefValueRight");
     gainStability             = configParam.getDouble("CCD/Gain/Stability");
     gainAllowedDifference     = configParam.getDouble("CCD/Gain/AllowedDifference");
+
+    readoutMode               = configParam.getString("CCD/ReadoutMode/ReadoutMode");
+
+    if(readoutMode == "Partial")
+    {
+    	firstRowPartialReadout = configParam.getInteger("CCD/ReadoutMode/Partial/FirstRowReadout");
+    	numRowsPartialReadout = configParam.getInteger("CCD/ReadoutMode/Partial/NumRowsReadout");
+    }
+    else if (readoutMode != "Nominal")
+    {
+    	Log.error("Detector::configure(): Unkown readout mode specification in configuration file: "  + readoutMode);
+    	throw ConfigurationException("Detector: Unkown readout mode specification in configuration file");
+    }
+
+    serialTransferTime = configParam.getDouble("CCD/SerialTransferTime") * 1E-9;			  // [ns] -> [s]
+    parallelTransferTime = configParam.getDouble("CCD/ParallelTransferTime") * 1E-6;		  // [µs] -> [s]
+    parallelTransferTimeFast = configParam.getDouble("CCD/ParallelTransferTimeFast") * 1E-6;  // [µs] -> [s]
 
     CTImodel                   = configParam.getString("CCD/CTI/Model");
     if (CTImodel == "Simple")
@@ -381,7 +403,14 @@ void Detector::updateParameters(double time)
     numRowsPixelMap         = configParam.getInteger("SubField/NumRows");
     numColumnsPixelMap      = configParam.getInteger("SubField/NumColumns");
     numRowsBiasMap          = configParam.getInteger("SubField/NumBiasPrescanRows");
+    numColumnsBiasMap       = configParam.getInteger("SubField/NumBiasPrescanColumns");
     numRowsSmearingMap      = configParam.getInteger("SubField/NumSmearingOverscanRows");
+
+    if(readoutMode == "Partial")
+    {
+    	Log.info("No smearing map for partial readout");
+    	numRowsSmearingMap = 0;
+    }
 
     // Configuration parameters for the noise source random seeds
 
@@ -390,12 +419,126 @@ void Detector::updateParameters(double time)
     cosmicSeed              = configParam.getLong("RandomSeeds/CosmicSeed");
     darkSignalSeed          = configParam.getLong("RandomSeeds/DarkSignalSeed");
 
+
     // Get the sequential number of the very first exposure
 
     beginExposureNr         = configParam.getInteger("ObservingParameters/BeginExposureNr");
 
     numEdgePixels = 0;
  }
+
+
+
+
+
+
+
+// /**
+//  * \brief Configure the readout times for Detector object using the ConfigurationParameters
+//  *
+//  * \param configParam: Configuration parameters
+//  **/
+//void Detector::configureReadoutTime(ConfigurationParameters &configParam)
+//{
+//	// Both detector halves are read out simultaneously
+//	// -> columns read out by the FEE:
+//	// 		- half of the CCD
+//	// 		- serial pre-scan
+//	// 		- (serial over-scan)
+//
+//	int numColumnsReadout = numColumns / 2 + numColumnsBiasMap; // + numRowsSerialOverScan
+//
+//	// How many rows will be actually read out by the FEE?
+//	// 	- nominal mode: image area + parallel over-scan
+//	//      normal camera: image area = whole CCD
+//	//      fast camera: image area = lower half of the CCD
+//	//	- partial readout: configurable
+//	// The rest of the image area will be dumped
+//
+//	int numRowsReadout, numRowsDump;
+//
+//	// -----------
+//	// Fast camera
+//	// -----------
+//
+//	if (isFastCamera) {
+//
+//		// Move the upper half of the CCD down to the lower half, row-by-row
+//
+//		int numRowsFrameTransfer = numRows - firstRowExposed;
+//
+//		readoutTimeBeforeNextExposure = numRowsFrameTransfer
+//						* parallelTransferTimeFast;
+//
+//		// The actual readout of the lower half of the CCD (after frame transfer) is done
+//		// while the next exposure has already started
+//
+//		// Nominal mode
+//
+//		if (readoutMode == "Nominal")
+//		{
+//			numRowsReadout = firstRowExposed + numRowsSmearingMap;
+//			numRowsDump = 0;
+//
+//		}
+//
+//		// Partial readout
+//
+//		else if (readoutMode == "Partial")
+//		{
+//			numRowsReadout = numRowsPartialReadout;
+//			numRowsDump = firstRowExposed - numRowsReadout;
+//		}
+//
+//
+//		readoutTimeDuringNextExposure = numRowsDump * parallelTransferTimeFast
+//				+ numRowsReadout * (parallelTransferTime + numColumnsReadout * serialTransferTime);
+//	}
+//
+//	// -------------
+//	// Normal camera
+//	// -------------
+//
+//	else
+//	{
+//
+//		// Nominal mode (full-frame readout)
+//
+//		if (readoutMode == "Nominal")
+//		{
+//
+//			// Rows read out by the FEE:
+//			// 		- rows of image area
+//			// 		- parallel over-scan
+//
+//			numRowsReadout = numRows + numRowsSmearingMap;
+//
+//			// No rows dumped
+//
+//			numRowsDump = 0;
+//		}
+//
+//		// Partial readout
+//
+//		else if (readoutMode == "Partial") {
+//
+//			// Rows read out by the FEE: rows in the block (other rows in image area are dumped)
+//			// Note: no parallel over-scan
+//
+//			numRowsReadout = numRowsPartialReadout;
+//			numRowsDump = numRows - numRowsReadout;
+//
+//		}
+//
+//		readoutTimeBeforeNextExposure =
+//				numRowsReadout
+//						* (numColumnsReadout * serialTransferTime
+//								+ parallelTransferTime)
+//						+ numRowsDump * parallelTransferTimeFast;
+//
+//		readoutTimeDuringNextExposure = 0;
+//	}
+// }
 
 
 
@@ -451,7 +594,7 @@ double Detector::takeExposure(int exposureNr, double startTime, double exposureT
 
     // Advance the internal clock
 
-    internalTime += exposureTime + readoutTime;
+    internalTime += exposureTime + readoutTimeBeforeNextExposure;
 
     return internalTime;
 }
@@ -761,7 +904,10 @@ void Detector::addDarkSignal(float exposureTime)
 
     // Add dark signal to the pixel map
 
-    double darkSignalRef = darkCurrent * (exposureTime + readoutTime);
+    // When is dark current accumulated for the pixel map?
+	// 	-  exposure + readout
+
+    double darkSignalRef = darkCurrent * (exposureTime + readoutTimeBeforeNextExposure + readoutTimeDuringNextExposure);
     darkSignalDistribution = normal_distribution<double>(darkSignalRef, darkSignalRef * dsnu / 100.0);
 
 
@@ -785,7 +931,13 @@ void Detector::addDarkSignal(float exposureTime)
 
     // Add dark signal to the smearing map
 
-    darkSignalRef = darkCurrent * readoutTime;
+    // When is dark current accumulated for the smearing map?
+    // 	- normal camera, nominal mode: whole readout
+    //  - fast camera, nominal mode: readout during the next exposure
+    //  - partial readout: no smearing map
+
+	darkSignalRef = darkCurrent
+			* (isFastCamera ? readoutTimeDuringNextExposure : readoutTimeBeforeNextExposure);
     darkSignalDistribution = normal_distribution<double>(darkSignalRef, darkSignalRef * dsnu / 100.0);
 
     for(unsigned int row = 0; row < numRowsSmearingMap; row++)
@@ -804,24 +956,6 @@ void Detector::addDarkSignal(float exposureTime)
             smearingMap(row, column) += darkSignal;
         }
     }
-}
-
-
-
-
-
-
-
-
-
-
-
-/**
- * \brief Return the CCD readout time [s].
- */
-double Detector::getReadoutTime()
-{
-    return readoutTime;
 }
 
 
@@ -1141,10 +1275,10 @@ void Detector::addPhotonNoise()
  */
 void Detector::addCosmics(float exposureTime)
 {
-    cosmicHitRateDistribution     = poisson_distribution<long>(cosmicHitRate);                                       // [hits/cm^2/s]
+	cosmicHitRateDistribution     = poisson_distribution<long>(cosmicHitRate);                                       // [hits/cm^2/s]
     cosmicEntryColumnDistribution = uniform_real_distribution<double>(0, numColumnsPixelMap - 1);                    // []
     cosmicEntryAngleDistribution  = uniform_real_distribution<double>(0, 2 * PI);                                    // [rad]
-    cosmicTrailLengthDistribution = uniform_real_distribution<double>(cosmicTrailLength[0], cosmicTrailLength[1]);   // [pix]
+    cosmicTrailLengthDistribution = uniform_real_distribution<double>(cosmicTrailLength[0], cosmicTrailLength[1]);   // [pixel]
     cosmicIntensityDistribution   = uniform_real_distribution<double>(cosmicIntensity[0], cosmicIntensity[1]);       // [e-/hit]
 
     // Cosmics in the subfield
@@ -1152,7 +1286,7 @@ void Detector::addCosmics(float exposureTime)
     if (includeCosmicsInSubField)
     {
         Log.debug("Detector: adding cosmic hits to the sub-field");
-        addCosmics(exposureTime + readoutTime, pixelMap, numRowsPixelMap, numColumnsPixelMap, "image area");
+        addCosmics(exposureTime + readoutTimeBeforeNextExposure + readoutTimeDuringNextExposure, pixelMap, numRowsPixelMap, numColumnsPixelMap, "image area");
     }
 
     // Cosmics in the over-scan
@@ -1160,7 +1294,14 @@ void Detector::addCosmics(float exposureTime)
     if (includeCosmicsInSmearingMap)
     {
         Log.debug("Detector: adding cosmic hits to smearing map");
-        addCosmics(readoutTime, smearingMap, numRowsSmearingMap, numColumnsPixelMap, "smearing map");
+
+        if(isFastCamera)
+        {
+        	addCosmics(readoutTimeDuringNextExposure, smearingMap, numRowsSmearingMap, numColumnsPixelMap, "smearing map");
+        } else
+        {
+        	addCosmics(readoutTimeBeforeNextExposure, smearingMap, numRowsSmearingMap, numColumnsPixelMap, "smearing map");
+        }
     }
 
     // Cosmics in the pre-scan
@@ -1171,8 +1312,9 @@ void Detector::addCosmics(float exposureTime)
     {
         Log.debug("Detector: adding cosmic hits to bias map");
         cosmicTrailLengthDistribution = uniform_real_distribution<double>(0.0, 1.e-6);    // Only hot pixels, no trails
-        const double biasMapReadoutTime = readoutTime / numRows * numRowsBiasMap; 
-        addCosmics(biasMapReadoutTime, biasMap, numRowsBiasMap, numColumnsPixelMap, "bias map");
+        const double biasMapRowLifeTime = (numColumns / 2 + numColumnsBiasMap) * serialTransferTime + parallelTransferTime;
+        addCosmics(biasMapRowLifeTime, biasMapLeft, numRowsBiasMap, numColumnsBiasMap, "bias map (left half)");
+        addCosmics(biasMapRowLifeTime, biasMapRight, numRowsBiasMap, numColumnsBiasMap, "bias map (right half)");
     }
 }
 
@@ -1621,7 +1763,8 @@ void Detector::applyShort2013CTImodel()
 
     // Compute the time it takes to transfer 1 row during readout
 
-    const double chargeTransferTime = readoutTime / numRows;                                                 // t [s]
+    const double chargeTransferTime = parallelTransferTime;		// t[s]
+//    const double chargeTransferTime = readoutTime / numRows;                                                 // t [s]
 
     // Compute the thermal velocity of the electrons in the silicon
 
@@ -1672,7 +1815,7 @@ void Detector::applyShort2013CTImodel()
             // Update the number of occupied traps with the estimated number of captured electrons
 
             numberOfOccupiedTraps.row(k) += numberOfCapturedElectrons;
-            
+
             // Correct the number of occupied traps with the electrons that were released again during the charge transfer time.
 
             numberOfReleasedElectrons = numberOfOccupiedTraps.row(k) * (1-exp(-chargeTransferTime/releaseTime[k]));
@@ -1728,7 +1871,8 @@ void Detector::applyOpenShutterSmearing(float exposureTime)
 
     arma::Row<float> openShutterSmearing = arma::sum(pixelMap, 0);
 
-    float openShutterSmearingOutsideSubField = camera.getTotalSkyBackground() * (numRows - numRowsPixelMap + numRowsSmearingMap);
+    int numRowsExposedDuringReadout = numRows - firstRowExposed - numRowsPixelMap + numRowsSmearingMap;
+    float openShutterSmearingOutsideSubField = camera.getTotalSkyBackground() * numRowsExposedDuringReadout;
 
     // Apply all throughput efficiencies
 
@@ -1753,7 +1897,10 @@ void Detector::applyOpenShutterSmearing(float exposureTime)
 
     openShutterSmearing += openShutterSmearingOutsideSubField;
 
-    float factor = (readoutTime / exposureTime) / numRows;
+    // Fast camera: lower half of the CCD is shielded off -> no contribution to open-shutter smearing here
+
+    float factor = readoutTimeBeforeNextExposure / exposureTime / (numRows - firstRowExposed);
+
     openShutterSmearing *= factor;
 
     // Add the effect of the open-shutter smearing to the pixel map
@@ -1826,12 +1973,13 @@ void Detector::addReadoutNoise()
 
     // Add readout noise to the bias prescan map
 
-    for (unsigned int row = 0; row < numRowsBiasMap; row++)
+    for(unsigned int row = 0; row < numRowsBiasMap; row++)
     {
-        for (unsigned int column = 0; column < numColumnsPixelMap; column++)
-        {
-            biasMap(row, column) += readoutNoiseDistribution(readoutNoiseGenerator);
-        }
+    	for(unsigned int column = 0; column < numColumnsBiasMap; column++)
+    	{
+    		biasMapLeft(row, column) += readoutNoiseDistribution(readoutNoiseGenerator);
+    		biasMapRight(row, column) += readoutNoiseDistribution(readoutNoiseGenerator);
+    	}
     }
 
     // Add readout noise to the smearing overscan map
@@ -1886,7 +2034,8 @@ void Detector::applyQuantisation()
     // maps do not have fractional values.
 
     pixelMap = arma::floor(pixelMap);
-    biasMap = arma::floor(biasMap);
+    biasMapLeft = arma::floor(biasMapLeft);
+    biasMapRight = arma::floor(biasMapRight);
     smearingMap = arma::floor(smearingMap);
 
 
@@ -1949,13 +2098,11 @@ void Detector::applyGain()
     if(lastIndexSubFieldLeft >= numColumnsPixelMap - 1)      // Left ADC only
     {
         pixelMap *= combinedGainLeft;
-        biasMap *= combinedGainLeft;
         smearingMap *= combinedGainLeft;
     }
     else if(lastIndexSubFieldLeft < 0)                     // Right ADC only
     {
         pixelMap *= combinedGainRight;
-        biasMap *= combinedGainRight;
         smearingMap *= combinedGainRight;
     }
     else
@@ -1963,15 +2110,16 @@ void Detector::applyGain()
         // 0 -> lastIndexSubFieldLeft: left ADC
 
         pixelMap.submat(arma::span::all, arma::span(0, lastIndexSubFieldLeft)) *= combinedGainLeft;
-        biasMap.submat(arma::span::all, arma::span(0, lastIndexSubFieldLeft)) *= combinedGainLeft;
         smearingMap.submat(arma::span::all, arma::span(0, lastIndexSubFieldLeft)) *= combinedGainLeft;
 
         // lastIndexSubFieldLeft + 1 -> numColumnsSubPixelMap -1: right ADC
 
         pixelMap.submat(arma::span::all, arma::span(lastIndexSubFieldLeft, numColumnsPixelMap - 1)) *= combinedGainRight;
-        biasMap.submat(arma::span::all, arma::span(lastIndexSubFieldLeft, numColumnsPixelMap - 1)) *= combinedGainRight;
         smearingMap.submat(arma::span::all, arma::span(lastIndexSubFieldLeft, numColumnsPixelMap - 1)) *= combinedGainRight;
     }
+
+    biasMapLeft *= combinedGainLeft;
+    biasMapRight *= combinedGainRight;
 
     Log.info("Detector: gain of left part of CCD: " + to_string(combinedGainLeft));
     Log.info("Detector: gain of right part of CCD: " + to_string(combinedGainRight));
@@ -2007,7 +2155,8 @@ void Detector::addElectronicOffset()
     // Add the electronic offset to the pixel, bias register, and smearing maps
 
     pixelMap += offset;
-    biasMap += offset;
+    biasMapLeft += offset;
+    biasMapRight += offset;
     smearingMap += offset;
 }
 
@@ -2041,7 +2190,8 @@ void Detector::applyDigitalSaturation()
 
     // Top off the values in the bias register map
 
-    biasMap(arma::find(biasMap > digitalSaturationLimit)).fill(digitalSaturationLimit);
+    biasMapLeft(arma::find(biasMapLeft > digitalSaturationLimit)).fill(digitalSaturationLimit);
+    biasMapRight(arma::find(biasMapRight > digitalSaturationLimit)).fill(digitalSaturationLimit);
 
     // Top off the values in the smearing map
 
@@ -2340,7 +2490,8 @@ void Detector::initHDF5Groups()
     Log.debug("Detector: initialising HDF5 groups");
 
     hdf5File.createGroup("/Images");
-    hdf5File.createGroup("/BiasMaps");
+    hdf5File.createGroup("/BiasMapsLeft");
+    hdf5File.createGroup("/BiasMapsRight");
     hdf5File.createGroup("/SmearingMaps");
     hdf5File.createGroup("/Flatfield");
     hdf5File.createGroup("/ThroughputMaps");
@@ -2365,7 +2516,7 @@ void Detector::initHDF5Groups()
 
 void Detector::writePixelMapsToHDF5(int exposureNr)
 {
-    // Compose the image name
+	// Compose the image name
 
     stringstream myStream;
     myStream << "image" << setfill('0') << setw(6) << exposureNr;
@@ -2375,19 +2526,23 @@ void Detector::writePixelMapsToHDF5(int exposureNr)
 
     hdf5File.writeArray("/Images", imageName, pixelMap);
 
-    // Clear the string stream and compose the smearing map name
+    if(numRowsSmearingMap != 0)
+    {
+    	// Clear the string stream and compose the smearing map name
 
-    myStream.str(string());      // insert empty string
-    myStream.clear();            // clear eof bit
+    	myStream.str(string());      // insert empty string
+    	myStream.clear();            // clear eof bit
 
-    myStream << "smearingMap" << setfill('0') << setw(6) << exposureNr;
-    string smearingMapName = myStream.str();
+    	myStream << "smearingMap" << setfill('0') << setw(6) << exposureNr;
+    	string smearingMapName = myStream.str();
 
-    // Add the smearing map to the "SmearingMaps" group
 
-    hdf5File.writeArray("/SmearingMaps", smearingMapName, smearingMap);
+    	// Add the smearing map to the "SmearingMaps" group
 
-    // Clear the string stream and compose the bias map name
+    	hdf5File.writeArray("/SmearingMaps", smearingMapName, smearingMap);
+    }
+
+   // Clear the string stream and compose the bias map name
 
     myStream.str(string());      // insert empty string
     myStream.clear();            // clear eof bit
@@ -2397,7 +2552,8 @@ void Detector::writePixelMapsToHDF5(int exposureNr)
 
     // Add the bias map to the "BiasMaps" group
 
-    hdf5File.writeArray("/BiasMaps", biasMapName, biasMap);
+    hdf5File.writeArray("/BiasMapsLeft", biasMapName, biasMapLeft);
+    hdf5File.writeArray("/BiasMapsRight", biasMapName, biasMapRight);
 
     // Clear the string stream and compose the throughput map name
 
@@ -2407,7 +2563,7 @@ void Detector::writePixelMapsToHDF5(int exposureNr)
     myStream << "throughputMap" << setfill('0') << setw(6) << exposureNr;
     string throughputMapName = myStream.str();
 
-    // Add the bias map to the "BiasMaps" group
+    // Add the throughput map to the "ThroughputMaps" group
 
     hdf5File.writeArray("/ThroughputMaps", throughputMapName, throughputMap);
 }
@@ -2427,3 +2583,21 @@ double Detector::getTemperature()
     return temperatureGenerator.getNextTemperature(internalTime);
 }
 
+
+
+
+
+
+
+
+
+
+
+
+/**
+ * \brief Returns the duration of the readout before the next exposure can start [s].
+ */
+double Detector::getReadoutTimeBeforeNextExposure()
+{
+	return readoutTimeBeforeNextExposure;
+}